island: Move setup/island_config_options.py -> cc/setup/
