+++ monkey/monkey_island/cc/setup/config_setup.py
@@ -5,7 +5,7 @@ from monkey_island.cc.arg_parser import IslandCmdArgs
 from monkey_island.cc.environment import server_config_handler
 from monkey_island.cc.environment.utils import create_secure_directory
 from monkey_island.cc.server_utils.consts import DEFAULT_DATA_DIR, DEFAULT_SERVER_CONFIG_PATH
-from monkey_island.setup.island_config_options import IslandConfigOptions
+from monkey_island.cc.setup.island_config_options import IslandConfigOptions
 
 
 def setup_data_dir(island_args: IslandCmdArgs) -> Tuple[IslandConfigOptions, str]:
