+++ monkey/monkey_island/cc/setup/island_config_options.py
